services/iam-service/requirements.txt
# Framework
fastapi==0.95.1
uvicorn==0.22.0

# Database
sqlalchemy==2.0.13
alembic==1.11.1
asyncpg==0.27.0

# Security
pyjwt==2.7.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Validation
pydantic==1.10.7
email-validator==2.0.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
httpx==0.24.0

# Utilities
python-dotenv==1.0.0
